작업중...
페이지권한처리...

+++ FDProject/src/main/java/com/fdproject/controller/ApiCartController.java
@@ -4,6 +4,7 @@ package com.fdproject.controller;
 import java.security.Principal;
 
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -13,24 +14,36 @@ import com.fdproject.domain.DrugsCartDTO;
 import com.fdproject.domain.RecipesCartDTO;
 import com.fdproject.service.DrugService;
 import com.fdproject.service.RecipeService;
+import com.fdproject.util.GrammerUtils;
+import com.fdproject.util.UiUtils;
 
 import lombok.RequiredArgsConstructor;
 
 @Controller
 @RequestMapping("/cart")
 @RequiredArgsConstructor
-public class ApiCartController {
+public class ApiCartController extends UiUtils {
     private final DrugService drugService;
     private final RecipeService recipeService;
 
     @PostMapping("/addDrug.do")
-    public @ResponseBody boolean insert1(@ModelAttribute DrugsCartDTO cartDTO) {
+    public @ResponseBody boolean insert1(@ModelAttribute DrugsCartDTO cartDTO, Principal principal, Model model) {
 
-        return drugService.addMyDrug(cartDTO);
+        String id = principal.getName();
+        if (GrammerUtils.isStringEmpty(id) == true) {
+            return false;
+        }
+
+        return drugService.addMyDrug(id, cartDTO);
     }
 
     @PostMapping("/deleteDrug.do")
-    public @ResponseBody boolean delete1(@ModelAttribute DrugsCartDTO cartDTO) {
+    public @ResponseBody boolean delete1(@ModelAttribute DrugsCartDTO cartDTO, Principal principal) {
+
+        String id = principal.getName();
+        if (GrammerUtils.isStringEmpty(id) == true) {
+            return false;
+        }
 
         return drugService.deleteMyDrug(cartDTO);
     }
